many many many many changes
apologies for broken things.  most of the models in InputFiles should work.

so mainly:
underworld/_underworld.py :  thinned down to a few basic functions
underworld/_stgermain.py  :  deals with interacting with stgermain library at the lower level
underworld/dictionary.py  :  deals with generation & manipulation of a global python dictionary, which represent the uw model.

there are many interface changes.  the goal here as been to simplify where ever possible.  note that all the dictionary functions need to be accessed like this (for eg):

import underworld

theglobaldictguy = underworld.dictionary.GetDictionary()

perhaps a bit verbose, but gives some hierarchy to functions.  really we should convert everything in dictionary.py to a python class.  next step perhaps.

also some significant changes to underlying libUnderworld code.  the aim of this is to allow component creation and destruction on the fly. more to come.